change time types

+++ xmtp/migrations/2023-05-19-010424_messages/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS messages (
     id INTEGER PRIMARY KEY NOT NULL,
-    created_at REAL NOT NULL,
+    created_at INTEGER NOT NULL,
     convoid TEXT NOT NULL,
     addr_from TEXT NOT NULL,
     content TEXT NOT NULL
